Added helper functions to get the most general type from a list of types, including Union types
