+++ src/types/ClassType.spec.ts
@@ -6,6 +6,7 @@ import { expectTypeToBe } from '../testHelpers.spec';
 import { ReferenceType } from './ReferenceType';
 import { isReferenceType } from '../astUtils/reflection';
 import { IntegerType } from './IntegerType';
+import { getUniqueType } from './helpers';
 
 describe('ClassType', () => {
 
@@ -41,7 +42,7 @@ describe('ClassType', () => {
         const superKlass = new ClassType('SuperKlass');
         superKlass.addMember('title', null, StringType.instance, SymbolTypeFlags.runtime);
         const subKlass = new ClassType('SubKlass', superKlass);
-        expectTypeToBe(subKlass.getMemberType('title', SymbolTypeFlags.runtime), StringType);
+        expectTypeToBe(getUniqueType(subKlass.getMemberTypes('title', SymbolTypeFlags.runtime)), StringType);
     });
 
     it('allow ReferenceTypes as super classes', () => {
@@ -59,7 +60,7 @@ describe('ClassType', () => {
         const futureSuperKlass = new ReferenceType('SuperKlass', SymbolTypeFlags.typetime, () => myTable);
         const subKlass = new ClassType('SubKlass', futureSuperKlass);
         expect(subKlass.isResolvable()).to.be.false;
-        const futureTitleType = subKlass.getMemberType('title', SymbolTypeFlags.runtime);
+        const futureTitleType = getUniqueType(subKlass.getMemberTypes('title', SymbolTypeFlags.runtime));
         expect(isReferenceType(futureTitleType)).to.be.true;
         expect(futureTitleType.isResolvable()).to.be.false;
         const superKlass = new ClassType('SuperKlass');
