+++ src/types/InheritableType.ts
@@ -12,8 +12,8 @@ export abstract class InheritableType extends BscType {
         }
     }
 
-    getMemberType(name: string, flags: SymbolTypeFlags) {
-        return super.getMemberType(name, flags) ?? new ReferenceType(name, flags, () => this.memberTable);
+    getMemberTypes(name: string, flags: SymbolTypeFlags) {
+        return super.getMemberTypes(name, flags) ?? [new ReferenceType(name, flags, () => this.memberTable)];
     }
 
     public toString() {
